Update with uthash commit 179d4d2c09368956253b6ccb6237158fe65349f7

+++ src/uthash/utstring.h
@@ -21,7 +21,7 @@ NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-/* a dynamic string implementation using macros 
+/* a dynamic string implementation using macros
  */
 #ifndef UTSTRING_H
 #define UTSTRING_H
@@ -29,13 +29,14 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #define UTSTRING_VERSION 1.9.9
 
 #ifdef __GNUC__
-#define _UNUSED_ __attribute__ ((__unused__)) 
+#define _UNUSED_ __attribute__ ((__unused__))
 #else
-#define _UNUSED_ 
+#define _UNUSED_
 #endif
 
 #include <stdlib.h>
 #include <string.h>
+#include <stdio.h>
 #include <stdarg.h>
 #define oom() exit(-1)
 
@@ -154,8 +155,8 @@ _UNUSED_ static void utstring_printf(UT_string *s, const char *fmt, ...) {
  ******************************************************************************/
 /* Build KMP table from left to right. */
 _UNUSED_ static void _utstring_BuildTable(
-    const char *P_Needle, 
-    size_t P_NeedleLen, 
+    const char *P_Needle,
+    size_t P_NeedleLen,
     long *P_KMP_Table)
 {
     long i, j;
@@ -163,7 +164,7 @@ _UNUSED_ static void _utstring_BuildTable(
     i = 0;
     j = i - 1;
     P_KMP_Table[i] = j;
-    while (i < P_NeedleLen)
+    while (i < (long) P_NeedleLen)
     {
         while ( (j > -1) && (P_Needle[i] != P_Needle[j]) )
         {
@@ -171,7 +172,7 @@ _UNUSED_ static void _utstring_BuildTable(
         }
         i++;
         j++;
-        if (i < P_NeedleLen)
+        if (i < (long) P_NeedleLen)
         {
             if (P_Needle[i] == P_Needle[j])
             {
@@ -194,8 +195,8 @@ _UNUSED_ static void _utstring_BuildTable(
 
 /* Build KMP table from right to left. */
 _UNUSED_ static void _utstring_BuildTableR(
-    const char *P_Needle, 
-    size_t P_NeedleLen, 
+    const char *P_Needle,
+    size_t P_NeedleLen,
     long *P_KMP_Table)
 {
     long i, j;
@@ -205,7 +206,7 @@ _UNUSED_ static void _utstring_BuildTableR(
     P_KMP_Table[i + 1] = j;
     while (i >= 0)
     {
-        while ( (j < P_NeedleLen) && (P_Needle[i] != P_Needle[j]) )
+        while ( (j < (long) P_NeedleLen) && (P_Needle[i] != P_Needle[j]) )
         {
            j = P_KMP_Table[j + 1];
         }
@@ -234,10 +235,10 @@ _UNUSED_ static void _utstring_BuildTableR(
 
 /* Search data from left to right. ( Multiple search mode. ) */
 _UNUSED_ static long _utstring_find(
-    const char *P_Haystack, 
-    size_t P_HaystackLen, 
-    const char *P_Needle, 
-    size_t P_NeedleLen, 
+    const char *P_Haystack,
+    size_t P_HaystackLen,
+    const char *P_Needle,
+    size_t P_NeedleLen,
     long *P_KMP_Table)
 {
     long i, j;
@@ -267,10 +268,10 @@ _UNUSED_ static long _utstring_find(
 
 /* Search data from right to left. ( Multiple search mode. ) */
 _UNUSED_ static long _utstring_findR(
-    const char *P_Haystack, 
-    size_t P_HaystackLen, 
-    const char *P_Needle, 
-    size_t P_NeedleLen, 
+    const char *P_Haystack,
+    size_t P_HaystackLen,
+    const char *P_Needle,
+    size_t P_NeedleLen,
     long *P_KMP_Table)
 {
     long i, j;
@@ -301,9 +302,9 @@ _UNUSED_ static long _utstring_findR(
 
 /* Search data from left to right. ( One time search mode. ) */
 _UNUSED_ static long utstring_find(
-    UT_string *s, 
+    UT_string *s,
     long P_StartPosition,   /* Start from 0. -1 means last position. */
-    const char *P_Needle, 
+    const char *P_Needle,
     size_t P_NeedleLen)
 {
     long V_StartPosition;
@@ -320,17 +321,17 @@ _UNUSED_ static long utstring_find(
         V_StartPosition = P_StartPosition;
     }
     V_HaystackLen = s->i - V_StartPosition;
-    if ( (V_HaystackLen >= P_NeedleLen) && (P_NeedleLen > 0) )
+    if ( (V_HaystackLen >= (long) P_NeedleLen) && (P_NeedleLen > 0) )
     {
         V_KMP_Table = (long *)malloc(sizeof(long) * (P_NeedleLen + 1));
         if (V_KMP_Table != NULL)
         {
             _utstring_BuildTable(P_Needle, P_NeedleLen, V_KMP_Table);
 
-            V_FindPosition = _utstring_find(s->d + V_StartPosition, 
-                                            V_HaystackLen, 
-                                            P_Needle, 
-                                            P_NeedleLen, 
+            V_FindPosition = _utstring_find(s->d + V_StartPosition,
+                                            V_HaystackLen,
+                                            P_Needle,
+                                            P_NeedleLen,
                                             V_KMP_Table);
             if (V_FindPosition >= 0)
             {
@@ -347,9 +348,9 @@ _UNUSED_ static long utstring_find(
 
 /* Search data from right to left. ( One time search mode. ) */
 _UNUSED_ static long utstring_findR(
-    UT_string *s, 
+    UT_string *s,
     long P_StartPosition,   /* Start from 0. -1 means last position. */
-    const char *P_Needle, 
+    const char *P_Needle,
     size_t P_NeedleLen)
 {
     long V_StartPosition;
@@ -366,17 +367,17 @@ _UNUSED_ static long utstring_findR(
         V_StartPosition = P_StartPosition;
     }
     V_HaystackLen = V_StartPosition + 1;
-    if ( (V_HaystackLen >= P_NeedleLen) && (P_NeedleLen > 0) )
+    if ( (V_HaystackLen >= (long) P_NeedleLen) && (P_NeedleLen > 0) )
     {
         V_KMP_Table = (long *)malloc(sizeof(long) * (P_NeedleLen + 1));
         if (V_KMP_Table != NULL)
         {
             _utstring_BuildTableR(P_Needle, P_NeedleLen, V_KMP_Table);
 
-            V_FindPosition = _utstring_findR(s->d, 
-                                             V_HaystackLen, 
-                                             P_Needle, 
-                                             P_NeedleLen, 
+            V_FindPosition = _utstring_findR(s->d,
+                                             V_HaystackLen,
+                                             P_Needle,
+                                             P_NeedleLen,
                                              V_KMP_Table);
 
             free(V_KMP_Table);
